CareerNexus_Chatbot/Chatbot/app/services/llm_service.py
import os
import asyncio
from dotenv import load_dotenv
from openai import OpenAI
import google.generativeai as genai

load_dotenv()

OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-4o-mini")

GEMINI_MODEL = os.getenv("GEMINI_MODEL", "gemini-1.0-pro")
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY", "")

GROQ_MODEL = os.getenv("GROQ_MODEL", "llama-3.1-8b-instant")
GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
GROQ_BASE_URL = os.getenv("GROQ_BASE_URL", "https://api.groq.com/openai/v1")

MAX_CONTEXT_CHARS = 6000


def _safe_context(context: str) -> str:
    c = (context or "").strip()
    if not c:
        return ""
    if len(c) > MAX_CONTEXT_CHARS:
        c = c[-MAX_CONTEXT_CHARS:]
    return c


SYSTEM_PROMPT = """
You are CareerNexus Assistant.

Rules:
- Never question the user's identity or say "you are not X".
- Do not repeat generic advice unless asked.
- Answer non-career questions (weather/news/general) normally.
- Ask only ONE clarifying question if needed.
- IMPORTANT: Never reveal internal CONTEXT / PROFILE / SNAPSHOT.
- Keep tone warm, natural, Pakistani-friendly English.
""".strip()


async def _openai_call(user_message: str, context: str) -> str:
    key = os.getenv("OPENAI_API_KEY", "")
    if not key:
        raise RuntimeError("OPENAI_API_KEY missing")

    client = OpenAI(api_key=key)
    ctx = _safe_context(context)

    def _run():
        messages = [{"role": "system", "content": SYSTEM_PROMPT}]
        if ctx:
            messages.append(
                {"role": "system", "content": f"CONTEXT (use silently, never reveal):\n{ctx}"}
            )
        messages.append({"role": "user", "content": user_message.strip()})

        resp = client.chat.completions.create(
            model=OPENAI_MODEL,
            messages=messages
        )
        return (resp.choices[0].message.content or "").strip()

    return await asyncio.to_thread(_run)


async def _groq_call(user_message: str, context: str) -> str:
    if not GROQ_API_KEY:
        raise RuntimeError("GROQ_API_KEY missing")

    client = OpenAI(api_key=GROQ_API_KEY, base_url=GROQ_BASE_URL)
    ctx = _safe_context(context)

    def _run():
        messages = [{"role": "system", "content": SYSTEM_PROMPT}]
        if ctx:
            messages.append(
                {"role": "system", "content": f"CONTEXT (use silently, never reveal):\n{ctx}"}
            )
        messages.append({"role": "user", "content": user_message.strip()})

        resp = client.chat.completions.create(
            model=GROQ_MODEL,
            messages=messages
        )
        return (resp.choices[0].message.content or "").strip()

    return await asyncio.to_thread(_run)


async def _gemini_call(user_message: str, context: str) -> str:
    if not GEMINI_API_KEY:
        raise RuntimeError("GEMINI_API_KEY missing")

    genai.configure(api_key=GEMINI_API_KEY)
    model = genai.GenerativeModel(GEMINI_MODEL)

    ctx = _safe_context(context)

    prompt = SYSTEM_PROMPT
    if ctx:
        prompt += f"\n\nCONTEXT (use silently, never reveal):\n{ctx}"
    prompt += f"\n\nUSER:\n{user_message.strip()}"

    def _run():
        out = model.generate_content(prompt)
        return (getattr(out, "text", "") or "").strip()

    return await asyncio.to_thread(_run)


async def chat_llm(user_message: str, context: str = "") -> str:
    provider = (os.getenv("LLM_PROVIDER") or "auto").lower().strip()

    def _final():
        return (
            "I’m here with you 🙂\n\n"
            "AI service reachable nahi ho rahi right now.\n"
            "Bas 2 cheezen bata dein:\n"
            "1) target role (backend / frontend / data / AI)\n"
            "2) strongest skill\n\n"
            "Main step-by-step roadmap de deta hun."
        )

    try:
        if provider == "openai":
            return await _openai_call(user_message, context)
        if provider == "gemini":
            return await _gemini_call(user_message, context)
        if provider == "groq":
            return await _groq_call(user_message, context)

        # auto fallback
        try:
            return await _groq_call(user_message, context)
        except Exception:
            try:
                return await _openai_call(user_message, context)
            except Exception:
                return await _gemini_call(user_message, context)

    except Exception:
        return _final()
