CareerNexus_Chatbot/Chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.12.5

sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0

python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.20

openai==1.43.0
google-generativeai==0.5.4
